+++ src/tokenize.rs
@@ -76,7 +76,7 @@ pub fn tokenize(path: String, input: String) -> Vec<Token> {
             // 初始化，类似于C++的构造函数
             // 我们不使用Head来存储信息，仅用来表示链表入口，这样每次都是存储在Cur->Next
             // 否则下述操作将使第一个Token的地址不在Head中。
-            let val = strtol(&chars, &mut pos, 10) as i32;
+            let val = strtol(&chars, &mut pos, 10);
             let t_str = slice_to_string(&chars, old_pos, pos);
             let t = Token::Num {
                 val,
